backend/requirements.txt
# FastAPI and async support
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ORM and database
tortoise-orm==0.21.6
asyncpg==0.29.0

# Environment and configuration
python-dotenv==1.0.1
pydantic-settings==2.5.2

# Additional dependencies
pydantic==2.9.2
